API

Creating plans

NonuniformFFTs.PlanNUFFT — Type

PlanNUFFT([T = ComplexF64], dims::Dims; ntransforms = Val(1), kwargs...)

Construct a plan for performing non-uniform FFTs (NUFFTs).

The created plan contains all data needed to perform NUFFTs for non-uniform data of type T (ComplexF64 by default) and uniform data with dimensions dims.

Optional keyword arguments

ntransforms = Val(1): the number of simultaneous transforms to perform. This is useful if one wants to transform multiple scalar quantities at the same non-uniform points.

NUFFT parameters

m = HalfSupport(8): the half-support of the convolution kernels. Large values increase accuracy at the cost of performance.

σ = 2.0: NUFFT oversampling factor. Typical values are 2.0 (more accurate) and 1.25 (faster), but other values such as 1.5 should also work.

kernel::AbstractKernel = BackwardsKaiserBesselKernel(): convolution kernel used for NUFFTs.

Performance parameters

block_size = 4096: the linear block size (in number of elements) when using block partitioning. This can be tuned for maximal performance. Using block partitioning is required for running with multiple threads. Blocking can be completely disabled by passing block_size = nothing (but this is generally slower, even when running on a single thread).

sort_points = False(): whether to internally permute the order of the non-uniform points. This can be enabled by passing sort_points = True(). In this case, more time will be spent in set_points! and less time on the actual transforms. This can improve performance if executing multiple transforms on the same non-uniform points. Note that, even when enabled, this does not modify the points argument passed to set_points!.

Other parameters

fftw_flags = FFTW.MEASURE: parameters passed to the FFTW planner.

timer = TimerOutput(): allows to specify a TimerOutput (from the TimerOutputs.jl package) where timing information will be written to. By default the plan creates its own timer. One can visualise the time spent on different parts of the NUFFT computation using p.timer.

Using real non-uniform data

In some applications, the non-uniform data to be transformed is purely real. In this case, one may pass Float64 or Float32 as the first argument. This may be faster than converting data to complex types, in particular because the real-to-complex FFTs from FFTW will be used to compute the transforms. Note that, in this case, the dimensions of the uniform data arrays is not exactly dims, since the size of the first dimension is divided roughly by 2 (taking advantage of Hermitian symmetry). For convenience, one can call size(::PlanNUFFT) on the constructed plan to know in advance the dimensions of the uniform data arrays.

Setting non-uniform points

NonuniformFFTs.set_points! — Function

set_points!(p::PlanNUFFT, points)

Set non-uniform points before executing a NUFFT.

In one dimension, points is simply a vector of real values (the non-uniform locations).

In multiple dimensions, points may be passed as:

a tuple of vectors (xs::AbstractVector, ys::AbstractVector, …),

a vector [x⃗₁, x⃗₂, x⃗₃, x⃗₄, …] of tuples or static vectors (typically SVectors from the StaticArrays.jl package).

The points are allowed to be outside of the periodic cell $[0, 2π)^d$, in which case they will be "folded" to that domain.

Executing plans

NonuniformFFTs.exec_type1! — Function

exec_type1!(ûs::AbstractArray{<:Complex}, p::PlanNUFFT, vp::AbstractVector{<:Number})
exec_type1!(ûs::NTuple{N, AbstractArray{<:Complex}}, p::PlanNUFFT, vp::NTuple{N, AbstractVector{<:Number}})

Perform type-1 NUFFT (from non-uniform points to uniform grid).

Here vp contains the input values at non-uniform points. The result of the transform is written into ûs.

One first needs to set the non-uniform points using set_points!.

To perform multiple transforms at once, both vp and ûs should be a tuple of arrays (second variant above). Note that this requires a plan initialised with ntransforms = Val(N) (see PlanNUFFT).

See also exec_type2!.

NonuniformFFTs.exec_type2! — Function

exec_type2!(vp::AbstractVector{<:Number}, p::PlanNUFFT, ûs::AbstractArray{<:Complex})
exec_type2!(vp::NTuple{N, AbstractVector{<:Number}}, p::PlanNUFFT, ûs::NTuple{N, AbstractArray{<:Complex}})

Perform type-2 NUFFT (from uniform grid to non-uniform points).

Here ûs contains the input coefficients in the uniform grid. The result of the transform at non-uniform points is written into vp.

One first needs to set the non-uniform points using set_points!.

To perform multiple transforms at once, both vp and ûs should be a tuple of arrays (second variant above). Note that this requires a plan initialised with ntransforms = Val(N) (see PlanNUFFT).

See also exec_type1!.

Other functions

Base.size — Method

size(p::PlanNUFFT) -> (N₁, N₂, ...)

Return the dimensions of arrays containing uniform values.

This corresponds to the number of Fourier modes in each direction (in the non-oversampled grid).

Index

NonuniformFFTs.PlanNUFFT

Base.size

NonuniformFFTs.exec_type1!

NonuniformFFTs.exec_type2!

NonuniformFFTs.set_points!